+++ src/main/java/org/smoothbuild/cli/taskmatcher/TaskMatchers.java
@@ -30,12 +30,13 @@ public class TaskMatchers {
   static final TaskMatcher CONVERSION = kindMatcher(TaskKind.CONVERSION);
   static final TaskMatcher FIELD = kindMatcher(TaskKind.FIELD_READ);
   static final TaskMatcher LITERAL = kindMatcher(TaskKind.LITERAL);
+  static final TaskMatcher MAP = kindMatcher(TaskKind.MAP);
   static final TaskMatcher REFERENCE = kindMatcher(TaskKind.REFERENCE);
   static final TaskMatcher VALUE = kindMatcher(TaskKind.VALUE);
 
   static final TaskMatcher DEFAULT = or(AT_LEAST_INFO, and(PRJ, or(or(CALL, VALUE), FIELD)));
 
-  private static final ImmutableMap<String, TaskMatcher> MAP =
+  private static final ImmutableMap<String, TaskMatcher> MATCHERS_MAP =
       ImmutableMap.<String, TaskMatcher>builder()
           .put("all", ALL)
           .put("a", ALL)
@@ -67,6 +68,8 @@ public class TaskMatchers {
           .put("field", FIELD)
           .put("literal", LITERAL)
           .put("l", LITERAL)
+          .put("map", MAP)
+          .put("m", MAP)
           .put("r", REFERENCE)
           .put("reference", REFERENCE)
           .put("value", VALUE)
@@ -75,7 +78,7 @@ public class TaskMatchers {
           .build();
 
   public static Optional<TaskMatcher> findMatcher(String name) {
-    return Optional.ofNullable(MAP.get(name));
+    return Optional.ofNullable(MATCHERS_MAP.get(name));
   }
 
   public static TaskMatcher and(TaskMatcher left, TaskMatcher right) {