+++ src/main/java/org/smoothbuild/lang/base/type/ItemSignature.java
@@ -23,6 +23,10 @@ public record ItemSignature(Type type, Optional<String> name, Optional<Type> def
     this.defaultValueType = requireNonNull(defaultValueType);
   }
 
+  public static ItemSignature itemSignature(Type type) {
+    return new ItemSignature(type, Optional.empty(), Optional.empty());
+  }
+
   public boolean hasDefaultValue() {
     return defaultValueType.isPresent();
   }